Vastly improve worker builds/bundles
Worker builds no longer come with the kitchen sink of languages and
themes.  Took a bit of a rework, I needed to break apart all the logic
for managing languages and themes to ensure the bundled stuff wouldn't
be included in the builds
