+++ packages/diffs/src/highlighter/languages/cleanUpResolvedLanguages.ts
@@ -1,0 +1,6 @@
+import { AttachedLanguages, ResolvedLanguages } from './constants';
+
+export function cleanUpResolvedLanguages(): void {
+  ResolvedLanguages.clear();
+  AttachedLanguages.clear();
+}
