+++ packages/diffs/src/highlighter/themes/cleanUpResolvedThemes.ts
@@ -1,0 +1,6 @@
+import { AttachedThemes, ResolvedThemes } from './constants';
+
+export function cleanUpResolvedThemes(): void {
+  ResolvedThemes.clear();
+  AttachedThemes.clear();
+}
