+++ packages/diffs/src/highlighter/themes/hasResolvedThemes.ts
@@ -1,0 +1,11 @@
+import type { DiffsThemeNames } from '../../types';
+import { ResolvedThemes } from './constants';
+
+export function hasResolvedThemes(themeNames: DiffsThemeNames[]): boolean {
+  for (const themeName of themeNames) {
+    if (!ResolvedThemes.has(themeName)) {
+      return false;
+    }
+  }
+  return true;
+}
