+++ packages/diffs/src/index.ts
@@ -2,9 +2,26 @@ export * from './components/File';
 export * from './components/FileDiff';
 export * from './components/FileStream';
 export * from './constants';
-export * from './highlighter/languages';
+export * from './highlighter/languages/areLanguagesAttached';
+export * from './highlighter/languages/attachResolvedLanguages';
+export * from './highlighter/languages/cleanUpResolvedLanguages';
+export * from './highlighter/languages/constants';
+export * from './highlighter/languages/getResolvedLanguages';
+export * from './highlighter/languages/getResolvedOrResolveLanguage';
+export * from './highlighter/languages/hasResolvedLanguages';
+export * from './highlighter/languages/resolveLanguage';
+export * from './highlighter/languages/resolveLanguages';
 export * from './highlighter/shared_highlighter';
-export * from './highlighter/themes';
+export * from './highlighter/themes/areThemesAttached';
+export * from './highlighter/themes/attachResolvedThemes';
+export * from './highlighter/themes/cleanUpResolvedThemes';
+export * from './highlighter/themes/constants';
+export * from './highlighter/themes/getResolvedOrResolveTheme';
+export * from './highlighter/themes/getResolvedThemes';
+export * from './highlighter/themes/hasResolvedThemes';
+export * from './highlighter/themes/registerCustomTheme';
+export * from './highlighter/themes/resolveTheme';
+export * from './highlighter/themes/resolveThemes';
 export * from './managers/LineSelectionManager';
 export * from './managers/MouseEventManager';
 export * from './managers/ResizeManager';
