+++ packages/diffs/src/renderers/FileRenderer.ts
@@ -3,12 +3,13 @@ import type { ElementContent, Element as HASTElement } from 'hast';
 import { toHtml } from 'hast-util-to-html';
 
 import { DEFAULT_THEMES } from '../constants';
-import { areLanguagesAttached } from '../highlighter/languages';
+import { areLanguagesAttached } from '../highlighter/languages/areLanguagesAttached';
 import {
   getHighlighterIfLoaded,
   getSharedHighlighter,
 } from '../highlighter/shared_highlighter';
-import { areThemesAttached, hasResolvedThemes } from '../highlighter/themes';
+import { areThemesAttached } from '../highlighter/themes/areThemesAttached';
+import { hasResolvedThemes } from '../highlighter/themes/hasResolvedThemes';
 import type {
   BaseCodeOptions,
   DiffsHighlighter,
